requirements.txt
torch
torchvision
requests
Pillow
pyttsx3
gradio